:sparkles: (modules): Create Home Assistant module
This module creates a Home Assistant container using podman. It also sets up a firewall rule to allow access to the Home Assistant web interface. The container is configured to use the host network and to access a USB device. The configuration is stored in a volume on the host.